Bug 1727146 - Only clear the preflight cache on user request. r=johannh
Differential Revision: https://phabricator.services.mozilla.com/D123134

UltraBlame original commit: 1d4ddeeef588b63b42cfff25da2b4d3e44eb531f

+++ toolkit/components/cleardata/ClearDataService.jsm
@@ -1304,12 +1304,21 @@ const AboutHomeStartupCacheCleaner = {
 };
 
 const PreflightCacheCleaner = {
-  deleteByPrincipal(aPrincipal) {
-    return this.deleteAll();
+  
+  
+  async deleteByPrincipal(aPrincipal, aIsUserRequest) {
+    if (!aIsUserRequest) {
+      return;
+    }
+    await this.deleteAll();
   },
 
-  deleteByBaseDomain(aBaseDomain) {
-    return this.deleteAll();
+  
+  async deleteByBaseDomain(aBaseDomain, aIsUserRequest) {
+    if (!aIsUserRequest) {
+      return;
+    }
+    await this.deleteAll();
   },
 
   async deleteAll() {
